requirements.txt
faster-whisper
edge-tts
sounddevice
numpy
keyboard
pystray
Pillow
miniaudio
pyttsx3
